refactor(smb): split create by replay, path resolution, and completion
